Finish extra credit component of Assignment 2:
(1.) Implement player vs. computer mode. In this version of the game, the player races against the computer to identify sets. The computer waits a random amount of time before removing a set from the board. If the player identifies a set first, the computer's timer resets. Whoever identifies the most sets wins. Moreover, expanding the number of cards on the board reduces the amount of time it takes for the computer to find a set, so there is a built-in penalty for doing so.

(2.) Implement a bonus for identifying sets quickly in one-player mode.